annotation-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.1
pandas==2.2.0
bcrypt==4.0.1
scipy==1.12.0
numpy==1.26.4
pytest==8.1.1
pytest-cov==4.1.0
httpx==0.27.0
